move assertion helpers into support directory
