+++ test/test_helper.rb
@@ -15,4 +15,7 @@ end
 
 require 'minitest/autorun'
 
+Dir[File.expand_path('../support/**/*.rb', __FILE__)].each { |f| require f }
+::MiniTest::Unit::TestCase.send(:include, AssertHelpers)
+
 require 'trollop'
