+++ test/test_trollop.rb
@@ -983,13 +983,6 @@ Options:
     assert_equal @q.leftovers, []
   end
 
-  def assert_parses_correctly(parser, commandline, expected_opts,
-                              expected_leftovers)
-    opts = parser.parse commandline
-    assert_equal expected_opts, opts
-    assert_equal expected_leftovers, parser.leftovers
-  end
-
   def test_unknown_subcommand
     @p.opt :global_flag, "Global flag", :short => "-g", :type => :flag
     @p.opt :global_param, "Global parameter", :short => "-p", :default => 5
@@ -1295,24 +1288,6 @@ Options:
       @p.opt :cb1, "with callback", :callback => lambda { |vals| raise "good" }
       @p.parse(%w(--cb1))
     end
-  end
-
-  private
-
-  def assert_stderr(msg = nil)
-    old_stderr, $stderr = $stderr, StringIO.new('')
-    yield
-    assert_match msg, $stderr.string if msg
-  ensure
-    $stderr = old_stderr
-  end
-
-  def assert_stdout(msg = nil)
-    old_stdout, $stdout = $stdout, StringIO.new('')
-    yield
-    assert_match msg, $stdout.string if msg
-  ensure
-    $stdout = old_stdout
   end
 end
